Add additional argument to Sender contract functions for handling wrapped tokens
Extended the Sender contract to handle wrapped tokens in transfers by adding an extra parameter in related methods. This modification allows for the use of both regular and permitted transfers of wrapped tokens. Necessary changes were also made to the contract constructor and GaslessSender.sol, and related scripts and tests were updated accordingly.