Add solhint disable comment in SenderUpgradeable contract
Added a solhint disable comment for the __Sender_init function in the SenderUpgradeable contract. This ensures that solhint, a linter for Solidity, does not flag this line of code for using mixed case for the function name. This mixed case naming is necessary to adhere to the upgradeable contract convention.